refactor: move search highlight implementation into SearchHighlighter
Add SearchHighlighter struct implementing Highlighter trait, which
implements search highlight logic (moved from HighlighterBundler).
