+++ src/editor/highlighter.rs
@@ -1,11 +1,9 @@
 mod rust;
+mod search_highlight;
 
-use super::{
-    annotated_string::{Annotation, Style},
-    buffer::Line,
-    RenderContext,
-};
+use super::{annotated_string::Annotation, buffer::Line, RenderContext};
 use crate::editor::filetype::FileType;
+use search_highlight::SearchHighlighter;
 
 trait Highlighter {
     fn highlight_line(&mut self, line: &Line);
@@ -21,45 +19,30 @@ fn create_syntax_highlighter(file_type: FileType) -> Option<Box<dyn Highlighter>
 
 pub struct HighlighterBundler<'a> {
     // line index to annotations of line.
-    highlights: Vec<Vec<Annotation>>,
     syntax_highlighter: Option<Box<dyn Highlighter>>,
-    render_context: &'a RenderContext,
+    search_highlighter: SearchHighlighter<'a>,
 }
 
 impl<'a> HighlighterBundler<'a> {
     pub fn new(context: &'a RenderContext) -> Self {
         Self {
-            highlights: Vec::new(),
             syntax_highlighter: create_syntax_highlighter(context.file_type),
-            render_context: context,
+            search_highlighter: SearchHighlighter::new(context),
         }
     }
 
     pub fn highlight_line(&mut self, line: &Line) {
-        let mut annotations = vec![];
         if self.syntax_highlighter.is_some() {
             self.syntax_highlighter
                 .as_mut()
                 .unwrap()
                 .highlight_line(line);
         }
-        self.highlight_search(line, &mut annotations);
-        self.highlights.push(annotations);
-    }
-
-    fn highlight_search(&self, line: &Line, annotations: &mut Vec<Annotation>) {
-        // search result annotations
-        let search_hits = match self.render_context.search_pattern.as_deref() {
-            Some(s) => line.search_all_occurence(s),
-            None => vec![],
-        };
-        for (match_start, match_end) in search_hits {
-            annotations.push(Annotation::new(Style::SearchHit, match_start, match_end));
-        }
+        self.search_highlighter.highlight_line(line);
     }
 
     pub fn get_annotations(&self, line_idx: usize) -> Vec<Annotation> {
-        let mut annotations = self.highlights[line_idx].clone();
+        let mut annotations = self.search_highlighter.get_annotations(line_idx);
         if self.syntax_highlighter.is_some() {
             annotations.append(
                 &mut self
